requirements.txt
fastapi==0.110.0
jose==1.0.0
numpy==1.24.3
passlib==1.7.4
pydantic==2.6.4
pytest==8.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.23
torch==2.2.1+cu118
transformers==4.39.0
